Give the room flags their own type, and Set[T] with them
docs/proposals/idiomatic-go.md's step 1, first domain of eleven, plus the
two pieces of shared machinery the other ten will lean on.

**What was wrong.** `internal/game/flags.go` defined one `Flags uint64`
and every flag in the game was that type — eleven unrelated domains
sharing not just a type but a namespace of *values*. `PlayerKiller`,
`PrefBrief` and `RoomDark` are all `1 << 0`, so
`rec.PlayerFlags.Has(RoomDark)` compiled, ran, and was true of every
killer in the game; `room.Flags.Has(PrefBrief)` likewise. Nothing in the
toolchain diagnoses that — not `go vet`, not `staticcheck`, not `gosec` —
because at the type level nothing is wrong. There is no evidence such a
mistake was in the tree; the argument is that if one were, nothing would
find it, and this port's own history (`isname`, wrong twice, the second
time with its oracle passing) is about exactly that class of thing.

**Set[T].** `internal/game/set.go`: a generic set over a per-domain enum,
with the domain's constants declared as bit *indices* rather than masks.
`Set[RoomFlag]` and `Set[PlayerFlag]` are unrelated types, so the mistake
above stops compiling. One generic type rather than eleven hand-written
ones, settled in the proposal's §4.1 before any of this was written:
eleven concrete types would read better at a call site and would also be
sixty-odd near-identical methods to keep in agreement, which is the
duplication §3.5 complains about, reinvented. This is the tree's first
generic *type* — there was one generic function, dump.go's `emptyIfNil` —
and that is a deliberate bet, not an oversight.

Nothing about the format moves. Bit 7 is still bit 7 on disk, is still
what `asciiflag_conv`'s letter decodes to, is still what `flags_raw`
carries, and is still what bitnames_test.go proves against
`room_bits[]` in constants.c. The `examples/` corpora are byte-identical,
which cmd/dlctl's import and resave fixtures both assert.

**The helper layer takes raw uint64 now.** `SprintBit`, `NameBits`,
`ParseBitNames` and the `asciiflag_conv` letter codec (`FlagLetters` /
`ParseFlagLetters`, split out of `Flags.String` / `ParseFlags`) all
operate on bits. Once every domain has its own set type there is no
single Go type left to write those signatures in, and the two
alternatives are worse: making each generic buys nothing, since none of
them can do anything with T that it cannot do with the bits, and it
would force the domain to be named at every call site; keeping eleven
copies is §3.5 again. So bits are the boundary, `Set.Raw`/`SetFromRaw`
are the only conversion, and everything above that layer is written in
the domain type. `Flags` and `ParseFlags` stay for the ten domains not
yet converted and go with the last of them.

**The trap this turned up, twice, in the first hour.** Bit indices mean
the C-shaped idiom still compiles and means something else:

    room.Flags.HasAny(RoomNoMob | RoomDeathTrap)   // asks about RoomIndoors
    room.Flags.HasAny(RoomNoMob, RoomDeathTrap)    // what was meant

`2 | 1` is `3`, a perfectly good `RoomFlag`. Types right, arity right,
answer silently about a different flag — the same "no diagnostic
anywhere" shape being retired, so trading one for the other would have
been a poor bargain. Both instances were real: `live.go`'s mobile
wandering (caught by a test somebody wrote years ago) and
`session/spells.go`'s summon check against private, death-trap and god
rooms, which nothing at all was watching.

Go cannot make this a type rule — any `~int` supports `|`, and the OR of
two valid indices is another valid index — so
`TestNoFlagConstantIsCombinedWithAnOperator` parses every Go file in the
tree with go/ast and rejects `|`, `&`, `&^` or `^` between two constants
of the same converted domain. It derives the constant list from the
`const` declarations rather than a hand-written list, so each domain a
later step converts is covered the moment it lands, with nothing to
remember. Verified by reintroducing one of the two bugs and watching it
fail.

Checks, per the proposal's §6: `make lint` 0, `go test ./...` green,
`make play` green (175s), `go test -race ./internal/server/` green
(181s). `make parity` is once per domain and this is one domain, but it
is also the first — running it before the next domain rather than
twice in a row.

+++ internal/game/identify.go
@@ -32,9 +32,9 @@ func IdentifyObject(obj *Object) string {
 
 	if obj.PermAffect != 0 {
 		fmt.Fprintf(&out, "Item will give you following abilities:  %s\r\n",
-			SprintBit(obj.PermAffect, affectBitNames))
+			SprintBit(uint64(obj.PermAffect), affectBitNames))
 	}
-	fmt.Fprintf(&out, "Item is: %s\r\n", SprintBit(obj.ExtraFlags, extraBitNames))
+	fmt.Fprintf(&out, "Item is: %s\r\n", SprintBit(uint64(obj.ExtraFlags), extraBitNames))
 	fmt.Fprintf(&out, "Weight: %d, Value: %d, Rent: %d, Min Level: %d\r\n",
 		obj.TotalWeight(), obj.Cost, obj.RentPerDay(), obj.MinLevel())
